impleamented the first version of practice screeen

+++ lib/Utils/Constants.dart
@@ -5,7 +5,7 @@ class Constants {
   static final double TITLE_FONT_SIZE = 15.0;
 }
 class ProgressBarConstants{
-    static final int PROGRESS_BAR_WIDTH = 300;
+    static final double PROGRESS_BAR_WIDTH = 300;
     static final int PROGRESS_BAR_TOTAL_STEPS = 100;
     static final Color PROGRESS_BAR_OUTLINE_COLOR = Colors.grey[400];
 }
